Trimming password to 15 characters
The sign-up and sign-in page both have an undocumented maximum length of 15 characters on the password. It's easy to type a longer password and not notice that the webpage will truncate it for you. Trying to log in with this tool would then result in an error, since you're sending a different password to the API. This patch resolves that issue by truncating the password before sending it in.

https://github.com/AHAAAAAAA/PokemonGo-Map/pull/85